\name{r_gpt}
\alias{r_gpt}
\title{R chatGPT copilot Function}
\description{
  This function allows the user to interact with the chat_gpt function in a more streamlined way and it is useful to obtain R code directly into the R session.
}
\usage{
  r_gpt(prompt, prompt0 = NULL, verbose = TRUE, n = 1)
}
\arguments{
  \item{prompt}{A character string representing the request for the user to write a script.}

  \item{prompt0}{An optional character string representing the initial prompt for the user. If not provided, the default prompt will be used.}

  \item{verbose}{A logical value indicating whether the function should print messages to the console during execution. The default value is \code{TRUE}.}

  \item{n}{An integer value indicating the number of responses to return. The default value is 1.}
}
\details{
  The \code{r_gpt} function is designed to help users with R programming tasks by prompting them to write scripts based on requests. The function uses the \code{chat_gpt} function to generate R code without explanations.
}
\value{
  The function returns the code generated by the \code{chat_gpt} function.
}
\examples{
  # Use the r_gpt function to generate code for a request.
  r_gpt("Create a histogram of the 'mpg' variable in the 'mtcars' dataset.")
}
